Move the worker code to its own module.
